+++ demo/vite.config.ts
@@ -1,14 +1,14 @@
-import { defineConfig } from 'vite'
+import { defineConfig } from "vite";
+import { resolve } from "path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  base: "/object-dom/",
   build: {
-    lib: {
-      entry: 'src/my-element.ts',
-      formats: ['es']
-    },
     rollupOptions: {
-      external: /^lit-element/
-    }
-  }
-})
+      input: {
+        main: resolve(__dirname, "index.html"),
+      },
+    },
+  },
+});