+++ src/dom/comment.ts
@@ -1,0 +1,4 @@
+export function createComment(content: string = "") {
+  const elem: Comment = document.createComment(content);
+  return elem;
+}