+++ src/dom/tags/iframe.ts
@@ -6,8 +6,15 @@ import { GlobalDom } from "../../object-dom";
  *
  * Reference: https://www.w3schools.com/TAGS/tag_iframe.asp
  */
-export class Iframe extends GlobalDom<HTMLElement> {
-  constructor(props: NodeProps<HTMLElement> = {}) {
+export class Iframe extends GlobalDom<HTMLIFrameElement> {
+  constructor(props: NodeProps<HTMLIFrameElement> = {}) {
     super({ node: document.createElement("iframe"), ...props });
+  }
+
+  replace(doc: Document) {
+    const current = this.node.contentDocument;
+    const srcNode = doc.documentElement;
+    const newNode = current!.importNode(srcNode, true);
+    current!.replaceChild(newNode, current!.documentElement);
   }
 }