+++ src/transformers/render.ts
@@ -1,5 +1,8 @@
 import type { ObjectDom } from "../base";
 
+/**
+ * Render the ObjectDom node at the given target.
+ */
 export function render(source: ObjectDom<HTMLElement>, target: HTMLElement = document.body) {
   let node = source.render().node;
   source.update = () => {