+++ src/transformers/text.ts
@@ -7,7 +7,7 @@ import type { GlobalDom } from "../object-dom";
  * Return all text content for a
  * given `ObjectDom` object
  */
-export function parseTextContent(source: GlobalDom<HTMLElement>): string {
+export function textContent(source: GlobalDom<HTMLElement>): string {
   const parser = new DOMParser();
   const doc = parser.parseFromString(generateHtml(source), "text/html");
   return doc.documentElement.textContent ?? "";