+++ types/dom/comment.d.ts
@@ -1,0 +1,1 @@
+export declare function createComment(content?: string): Comment;