+++ types/dom/index.d.ts
@@ -1,5 +1,7 @@
 export * from './attrs';
+export * from './comment';
 export * from './custom';
+export * from './document';
 export * from './styles';
 export * from './tags/a';
 export * from './tags/abbr';