+++ types/dom/tags/iframe.d.ts
@@ -5,6 +5,7 @@ import { GlobalDom } from "../../object-dom";
  *
  * Reference: https://www.w3schools.com/TAGS/tag_iframe.asp
  */
-export declare class Iframe extends GlobalDom<HTMLElement> {
-    constructor(props?: NodeProps<HTMLElement>);
+export declare class Iframe extends GlobalDom<HTMLIFrameElement> {
+    constructor(props?: NodeProps<HTMLIFrameElement>);
+    replace(doc: Document): void;
 }