+++ types/transformers/generate.d.ts
@@ -1,8 +1,8 @@
-import type { GlobalDom } from "../object-dom";
+import type { GlobalDom, ObjectDocument } from "../object-dom";
 /**
  * @returns string
  *
  * Return all text content for a
  * given `ObjectDom` object
  */
-export declare function generateHtml(source: GlobalDom<HTMLElement>): string;
+export declare function generateHtml(source: GlobalDom<HTMLElement> | ObjectDocument): string;