+++ types/transformers/render.d.ts
@@ -1,2 +1,5 @@
 import type { ObjectDom } from "../base";
+/**
+ * Render the ObjectDom node at the given target.
+ */
 export declare function render(source: ObjectDom<HTMLElement>, target?: HTMLElement): void;